% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImfDataGui.r
\name{ImfDataGui}
\alias{ImfDataGui}
\title{Retrieve data from IMF (IFS)}
\usage{
ImfDataGui(varname = "gdp", series = "Gross Domestic Product",
  startdate = "1980-01-01", enddate = "2016-12-31")
}
\arguments{
\item{varname}{Name to be given to output (panel)}

\item{series}{Keyword(s) to search IFS dataset}

\item{startdate}{Starting date of series}

\item{enddate}{End date of series}
}
\value{
panel data in data.frame format
}
\description{
Retrieve data from IMF (IFS)
}
\examples{
#' ImfDataGui(varname="gdp",series='Gross Domestic Product',startdate='1980-01-01',enddate='2016-12-31')

}
